Splits out POC code; adds utils.test
